Update Codename "Ogiso Susumu"
####2014Dec04(MST) Wave 4, Codename "Ogiso Susumu"####
These updates were introduced regarding to the last wave:

1. SHSDeskInterface updated, only its lineGap changed to 50 while other
metrics are reverted to Adobe's factorial settings (thanks to Kobayashi
Ken, a.k.a. Ken Lunde, the programmer of SHS, for his suggestion).

2. Since OS X Yosemite's stable release has already been 14B25, all
concerned files in this repo and the OSXCJKFontPlists repo has been
updated, incl. AEDKit Scripts.

3. Thanks to Wells Riley @
https://github.com/wellsriley/YosemiteSanFranciscoFont , his repo
inspired me that the system font could be installed in
"/Library/Fonts/" folder. This means that only `chown root:wheel` +
`chmod 644` permission settings and `sudo atsutil databases -remove`
are mandatory regarding to the installation of new system font files.
Thus, we updated this repo and the OSXCJKFontPlists repo.